requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-environ==0.9.0
django-filter==22.1

djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
